Break apart unit tests into individual concerns
